:green_heart: chore: test 폴더 내 store 패키지 생성 및 coin 패키지 이동
