+++ gg-pingpong-api/src/test/java/gg/pingpong/api/admin/store/controller/CoinAdminControllerTest.java
@@ -1,9 +1,10 @@
-package gg.pingpong.api.admin.coin.controller;
+package gg.pingpong.api.admin.store.controller;
 
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import org.apache.http.HttpHeaders;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,9 +15,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import gg.admin.repo.store.CoinPolicyAdminRepository;
 import gg.auth.utils.AuthTokenProvider;
-import gg.pingpong.api.admin.store.dto.CoinPolicyAdminAddDto;
+import gg.pingpong.api.admin.store.controller.request.CoinUpdateRequestDto;
+import gg.repo.user.UserRepository;
 import gg.utils.TestDataUtils;
 import gg.utils.annotation.IntegrationTest;
 import lombok.RequiredArgsConstructor;
@@ -25,7 +26,7 @@ import lombok.RequiredArgsConstructor;
 @IntegrationTest
 @AutoConfigureMockMvc
 @Transactional
-public class CoinPolicyAdminControllerFailTest {
+class CoinAdminControllerTest {
 	@Autowired
 	TestDataUtils testDataUtils;
 
@@ -39,24 +40,21 @@ public class CoinPolicyAdminControllerFailTest {
 	AuthTokenProvider tokenProvider;
 
 	@Autowired
-	CoinPolicyAdminRepository coinPolicyAdminRepository;
+	UserRepository userRepository;
 
 	@Test
-	@DisplayName("[Post FAIL]/pingpong/admin/coinpolicy")
-	void addAnnouncement() throws Exception {
+	@DisplayName("PUT /pingpong/admin/coin")
+	public void updateCoinTest() throws Exception {
 		String accessToken = testDataUtils.getAdminLoginAccessToken();
 		Long userId = tokenProvider.getUserIdFromAccessToken(accessToken);
-
-		CoinPolicyAdminAddDto addDto = new CoinPolicyAdminAddDto(1, 2, 5, -1);
-
-		String content = objectMapper.writeValueAsString(addDto);
-		String url = "/pingpong/admin/coinpolicy";
-
-		String contentAsString = mockMvc.perform(post(url)
-				.content(content)
+		String creatorId = userRepository.getById(userId).getIntraId();
+		int beforeCoin = userRepository.getById(userId).getGgCoin();
+		int changeCoin = 10;
+		CoinUpdateRequestDto coinUpdateRequestDto = new CoinUpdateRequestDto(creatorId, changeCoin, "관리자 코인 지급 테스트");
+		mockMvc.perform(put("/pingpong/admin/coin").header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
 				.contentType(MediaType.APPLICATION_JSON)
-				.header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken))
-			.andExpect(status().isBadRequest())
-			.andReturn().getResponse().getContentAsString();
+				.content(objectMapper.writeValueAsString(coinUpdateRequestDto)))
+			.andExpect(status().isNoContent());
+		Assertions.assertThat(userRepository.getById(userId).getGgCoin()).isEqualTo(beforeCoin + changeCoin);
 	}
 }
